moving some of the welcome activity monitors to their own file
