+++ commands/index.js
@@ -1,4 +1,5 @@
 module.exports = {
   config: require('./config.js'),
-  channel_activity: require('./channel_activity.js')
+  channel_activity: require('./channel_activity.js'),
+  welcome_activity: require('./welcome_activity.js')
 };
